+++ test/spec/guards/server.spec.ts
@@ -10,6 +10,8 @@ const http = pointy.http;
 let serverfork;
 
 beforeAll(async () => {
+	jasmine.DEFAULT_TIMEOUT_INTERVAL = 10000;
+
 	serverfork = await forkServer('./lib/test/examples/guards/server.js');
 
 	// Database